src/Services/Chatbot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
langchain==0.3.7
langchain-core==0.3.17
langchain-community==0.3.5
langchain-openai==0.2.8
langgraph==0.2.45
openai==1.54.3
PyJWT==2.8.0
